# api-spec.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: >-
      https://virtserver.swaggerhub.com/matthewminish/seng302-2021-api-spec/3.0.1
info:
  description: API specification for the SENG302 2021 project
  version: 3.0.1
  title: Wasteless
  contact:
    email: seng302@canterbury.ac.nz
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
security:
  - CookieAuth:
      - user
      - globalApplicationAdmin
      - defaultGlobalApplicationAdmin
paths:
...
  '/businesses/{id}/listings':
    get:
      tags:
        - 'U22: List sale'
        - Business Accounts
        - Listings / sales
      summary: Retrieve a business's listings
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
        - name: page
          in: query
          description: The page number to retrieve (1 indexed)
          schema:
            type: integer
        - name: resultsPerPage
          in: query
          description: The number of results to show per page
          schema:
            type: integer
        - name: orderBy
          in: query
          description: Card property to order results by
          schema:
            type: string
            enum:
              - created
              - closing
              - productCode
              - productName
              - quantity
              - price
        - name: reverse
          in: query
          description: Whether to flip default sort order (default is ascending)
          schema:
            type: boolean
      responses:
        '200':
          description: Listings retrieved successfully. This could be an empty array.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required:
                      - results
                    properties:
                      results:
                        type: array
                        items:
                          $ref: '#/components/schemas/Listing'

        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
    post:
      tags:
        - 'U22: List sale'
        - Business Accounts
        - Listings / sales
      summary: Add a new listing for some inventory item of a business
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: Business ID
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Listing created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  listingId:
                    type: integer
                    description: The ID of the listing that has just been created
                    example: 57
        '400':
          description: >-
            There was some error with the data supplied by the user, appropriate
            error message(s) should be shown to user
        '403':
          description: >-
            Forbidden: Returned when a user tries to add listing for a business
            they do not administer AND the user is not a global application
            admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Listing'
  '/listings/{id}/interest':
    put:
      tags:
        - 'U30: Individual full sale listing'
        - Listings / sales
      summary: Sets the interest state (liked/unliked) for the given user and listing
      security:
      - CookieAuth:
          - user
          - globalApplicationAdmin
          - defaultGlobalApplicationAdmin
      parameters:
        - name: id
          in: path
          description: ID of listing to update the interest state
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - interested
              properties:
                userId:
                  type: integer
                  example: 6
                  description: ID of user to update the interest state
                interested:
                  type: boolean
                  example: false
                  description: New interest state for the provided user/listing combination
      responses:
        '200':
          description: Request successful and interest state is updated
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '406':
          $ref: '#/components/responses/NotAcceptableError'
  /businesses/listings/search:
    get:
      tags:
        - 'U29: Search Sale Listing'
        - Business Accounts
        - Listings / sales
      summary: Get sale items for business that meet search criteria passed
      security:
        - CookieAuth:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
      parameters:
        - name: basicSearchQuery
          in: query
          description: string to match against any field of sale items
          schema:
            type: string
            example: "Cheese"
        - name: productSearchQuery
          in: query
          description: string to match against name or description of sale items
          schema:
            type: string
            example: "Ham"
        - name: businessSearchQuery
          in: query
          description: string to match against business names
          schema:
            type: string
            example: "John's Butcher"
        - name: locationSearchQuery
          in: query
          description: string to match against business addresses
          schema:
            type: string
            example: "23 Here St"
        - name: page
          in: query
          description: The page number to retrieve (1 indexed)
          schema:
            type: integer
            example: 1
        - name: resultsPerPage
          in: query
          description: The number of results to show per page
          schema:
            type: integer
            example: 10
        - name: orderBy
          in: query
          description: Sale property to order results by
          schema:
            type: string
            example: created
            enum:
              - name
              - price
              - businessName
              - businessLocation
              - closes
              - created
              - quantity
        - name: reverse
          in: query
          description: Whether to flip default sort order (default is ascending)
          schema:
            type: boolean
            example: false
        - name: businessType
          in: query
          description: Business types to filter results by
          schema:
            $ref: '#/components/schemas/BusinessTypes'
        - name: priceLower
          in: query
          description: Results must be above this price inclusive
          schema:
            type: number
            format: double
            example: 13.99
        - name: priceUpper
          in: query
          description: Results must be below this price inclusive
          schema:
            type: number
            format: double
            example: 15.50
        - name: closeLower
          in: query
          description: Results must close after this date
          schema:
            type: string
            format: date-time
            example: '2021-07-21T23:59:00Z'
        - name: closeUpper
          in: query
          description: Results must close before this date
          schema:
            type: string
            format: date-time
            example: '2021-07-21T23:59:00Z'
      responses:
        '200':
          description: 'Request successful, a (possibly empty) list of sale items is returned'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required:
                      - results
                    properties:
                      results:
                        type: array
                        items:
                          allOf:
                            - $ref: '#/components/schemas/Listing'
                            - type: object
                              required:
                                - business
                              properties:
                                business:
                                  $ref: '#/components/schemas/Business'
        '400':
          description: >-
            Invalid search term provided
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  responses:
    BadRequestError:
      description: >-
        There was some error with the data supplied by the user, appropriate
        error message(s) should be shown to user
    UnauthorizedError:
      description: Access token is missing or invalid
...
    NotAcceptableError:
      description: >-
        The requested route does exist (so not a 404) but some part of the
        request is not acceptable, for example trying to access a resource by an
        ID that does not exist.
...
  schemas:
...
    Listing:
      type: object
      description: A listing for the sale of some product from a business's inventory
      required:
        - inventoryItemId
        - quantity
        - price
      properties:
        id:
          type: integer
          example: 57
          readOnly: true
          description: >-
            A unique (within the business) identifier for this listing -
            autogenerated by the system
        inventoryItem:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/InventoryItem'
        inventoryItemId:
          type: integer
          example: 101
          writeOnly: true
        quantity:
          type: integer
          example: 3
          description: >-
            Must be greater than zero, and less than or equal to the quantity of
            the correpsonding inventory entry
        price:
          type: number
          format: double
          description: Price of the listing
          example: 17.99
        moreInfo:
          type: string
          example: Seller may be willing to consider near offers
          description: An optional field for further information about the listing
        created:
          type: string
          format: date-time
          example: '2021-07-14T11:44:00Z'
          readOnly: true
          description: Date and time when the listing was created
        closes:
          type: string
          format: date-time
          example: '2021-07-21T23:59:00Z'
          description: >-
            Date and time when the listing closes (defaults to expiry date of
            related inventory item)
    InventoryItem:
      type: object
      description: Entity representing an instance of a stock of some product
      required:
        - productId
        - quantity
        - remainingQuantity
        - expires
      properties:
        id:
          type: integer
          example: 101
          readOnly: true
          description: >-
            A unique (within the inventory) identifer for this item -
            autogenerated by the system
        product:
          readOnly: true
          allOf:
            - $ref: '#/components/schemas/Product'
        productId:
          writeOnly: true
          type: string
          example: WATT-420-BEANS
          description: >-
            Product Code of corresponding product in business's catalogue.
            Notice that the ID is given when writing, but the full object is
            returned when reading
        quantity:
          type: integer
          example: 4
          description: How many items of stock are present
        remainingQuantity:
          type: integer
          example: 3
          description: How many items of stock are not in a listing
        pricePerItem:
          type: number
          format: double
          description: Price to purchase a single item of the product
          example: 6.5
        totalPrice:
          type: number
          format: double
          description: Price for purchasing the entire stock of the product
          example: 21.99
        manufactured:
          type: string
          format: date
          description: The date on which the product was prepared / manufactured
        sellBy:
          type: string
          format: date
          description: The date by which the product ought to be sold
        bestBefore:
          type: string
          format: date
          description: The "best before" date of the product
        expires:
          type: string
          format: date
          description: The expiry date of the product
    Image:
      type: object
      description: Entity representing an image stored on the server
      required:
        - filename
        - thumbnailFilename
      properties:
        id:
          type: integer
          example: 1234
          readOnly: true
          description: A unique identifier (system-wide) to represent this image
        filename:
          type: string
          example: /media/images/23987192387509-123908794328.png
          description: >-
            Relative URI to full size image. A fully qualified URI (e.g
            www.wasteless.com/media/images/...) should NOT be given.
        thumbnailFilename:
          type: string
          example: /media/images/23987192387509-123908794328_thumbnail.png
          description: >-
            Relative URI to thumbnail of image. A fully qualified URI (e.g
            www.wasteless.com/media/images/...) should NOT be given.
...
    Business:
      type: object
      description: Business account entity
      required:
        - name
        - address
        - businessType
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        administrators:
          type: array
          items:
            $ref: '#/components/schemas/User'
          readOnly: true
        primaryAdministratorId:
          type: integer
          example: 20
          description: >-
            The ID of the 'primary' administrator of the business. This is
            automatically given to the creator of the business.
        name:
          type: string
          example: Lumbridge General Store
        description:
          type: string
          example: A one-stop shop for all your adventuring needs
        address:
          $ref: '#/components/schemas/Address'
        businessType:
          $ref: '#/components/schemas/BusinessType'
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:52:00Z'
          readOnly: true
          description: Date and time when the account was created
        images:
          type: array
          readOnly: true
          description: List of images for this business, with the first image being the primary image.
          items:
            $ref: '#/components/schemas/Image'
    BusinessType:
      type: string
      example: Retail Trade
      enum:
        - Accommodation and Food Services
        - Retail Trade
        - Charitable organisation
        - Non-profit organisation
    BusinessTypes:
      type: object
      required:
        - values
      properties:
        values:
          type: array
          items:
            $ref: '#/components/schemas/BusinessType'
    Address:
      type: object
      description: Entity representing some physical address
      required:
        - country
      properties:
        streetNumber:
          type: string
          example: 3/24
        streetName:
          type: string
          example: Ilam Road
        district:
          type: string
          example: Upper Riccarton
        city:
          type: string
          example: Christchurch
        region:
          type: string
          example: Canterbury
        country:
          type: string
          example: New Zealand
        postcode:
          type: string
          example: '90210'
    User:
      type: object
      description: Basic user entity
      required:
        - firstName
        - lastName
        - email
        - dateOfBirth
        - homeAddress
        - password
      properties:
        id:
          type: integer
          example: 100
          readOnly: true
          description: Automatically generated and assigned by the server
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Smith
        middleName:
          type: string
          example: Hector
        nickname:
          type: string
          example: Jonny
        bio:
          type: string
          example: Likes long walks on the beach
        email:
          type: string
          format: email
          example: johnsmith99@gmail.com
        dateOfBirth:
          type: string
          format: date
          example: '1999-04-27'
        phoneNumber:
          type: string
          example: +64 3 555 0129
        homeAddress:
          $ref: '#/components/schemas/Address'
        password:
          type: string
          example: 1337-H%nt3r2
          format: password
          writeOnly: true
          description: Should not be stored in plain text (i.e hash it)
        created:
          type: string
          format: date-time
          example: '2020-07-14T14:32:00Z'
          readOnly: true
          description: Date and time when the account was created
        role:
          type: string
          readOnly: true
          description: >-
            (Implemented in U4): This property should only be shown to Global
            Application Admins
          enum:
            - user
            - globalApplicationAdmin
            - defaultGlobalApplicationAdmin
        businessesAdministered:
          type: array
          readOnly: true
          description: >-
            (Implemented in U6): A list of businesses for which the user is an
            administrator
          items:
            $ref: '#/components/schemas/Business'
    Page:
      type: object
      description: Page of results
      required:
        - count
      properties:
        count:
          type: integer
          example: 70
          readOnly: true
          description: Total number of results
...